Remove default IOStream: SCStreamOutput extension.
