+++ Examples/macOS/SCStreamPublishViewController.swift
@@ -85,3 +85,19 @@ extension SCStreamPublishViewController: SCStreamDelegate {
         print(error)
     }
 }
+
+extension IOStream: SCStreamOutput {
+    @available(macOS 12.3, *)
+    public func stream(_ stream: SCStream, didOutputSampleBuffer sampleBuffer: CMSampleBuffer, of type: SCStreamOutputType) {
+        if #available(macOS 13.0, *) {
+            switch type {
+            case .screen:
+                append(sampleBuffer)
+            default:
+                append(sampleBuffer)
+            }
+        } else {
+            append(sampleBuffer)
+        }
+    }
+}
