+++ Sources/IO/IOStream.swift
@@ -4,9 +4,6 @@ import CoreMedia
 #if canImport(SwiftPMSupport)
 import SwiftPMSupport
 #endif
-#if canImport(ScreenCaptureKit)
-import ScreenCaptureKit
-#endif
 #if canImport(UIKit)
 import UIKit
 #endif
@@ -558,20 +555,3 @@ extension IOStream: IOScreenCaptureUnitDelegate {
     }
 }
 
-#if os(macOS)
-extension IOStream: SCStreamOutput {
-    @available(macOS 12.3, *)
-    public func stream(_ stream: SCStream, didOutputSampleBuffer sampleBuffer: CMSampleBuffer, of type: SCStreamOutputType) {
-        if #available(macOS 13.0, *) {
-            switch type {
-            case .screen:
-                append(sampleBuffer)
-            default:
-                append(sampleBuffer)
-            }
-        } else {
-            append(sampleBuffer)
-        }
-    }
-}
-#endif
